csbot/requirements.txt
pytube
discord
dadjokes
cachetools
python-dotenv
masterblaster.py